requirements.txt
python-dotenv==1.2.1
requests==2.32.5
flask==3.1.2